[Refactor/#191] 검색 컨트롤러 및 도메인별 검색 어댑터 구현 완료
